requirements.txt
Django==2.1
djangorestframework==3.8.2
djangorestframework-jsonapi==2.5.0
inflection==0.3.1
pytz==2018.5
six==1.11.0
